+++ tests/testthat/test-eval.R
@@ -1,4 +1,4 @@
-context("invoke")
+context("eval")
 
 test_that("invoke() buries arguments", {
   expect_identical(invoke(call_inspect, 1:2, 3L), quote(.fn(`1`, `2`, `3`)))
@@ -10,4 +10,12 @@ test_that("invoke() buries arguments", {
 test_that("invoke() can be called without arguments", {
   expect_identical(invoke("list"), list())
   expect_identical(invoke(list), list())
+})
+
+test_that("locally() supports multiple arguments", {
+  exprs <- exprs(foo = toupper("foo"), message("bar"), baz = paste(foo, "baz"), baz)
+  expect_message(expect_identical(locally(!!!exprs), "FOO baz"), "bar")
+
+  env <- locally(!!!exprs, current_env())
+  expect_true(all(env_names(env) %in% c("foo", "baz")))
 })